backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==3.14.0
django-jazzmin==3.0.1
django-shortuuidfield==0.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
Pygments==2.19.1
PyJWT==2.6.0
pytz==2025.2
PyYAML==6.0.2
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
